+++ src/main/java/org/example/plan/service/PlanService.java
@@ -4,8 +4,10 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.example.plan.config.PasswordEncoder;
 import org.example.plan.dto.PlanResponseDto;
+import org.example.plan.entity.Comment;
 import org.example.plan.entity.Member;
 import org.example.plan.entity.Plan;
+import org.example.plan.repository.CommentRepository;
 import org.example.plan.repository.MemberRepository;
 import org.example.plan.repository.PlanRepository;
 import org.springframework.http.HttpStatus;
@@ -23,6 +25,7 @@ public class PlanService {
     private final MemberRepository memberRepository;
     private final PlanRepository planRepository;
     private final PasswordEncoder passwordEncoder;
+    private final CommentRepository commentRepository;
 
     public PlanResponseDto save(String username, String title, String contents){
         Member findMember = memberRepository.findMemberByUsernameOrElseThrow(username);
@@ -50,13 +53,14 @@ public class PlanService {
 
     @Transactional
     public PlanResponseDto updatePlan(Long id,  String title, String contents, String password) {
-        Member member = memberRepository.findByIdOrElseThrow(id);
+
+        Plan findPlan = planRepository.findByIdOrElseThrow(id);
+        Member member = findPlan.getMember();
 
         if(!passwordEncoder.matches(password,member.getPassword())){
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST," 비밀번호 오답");
         }
 
-        Plan findPlan = planRepository.findByIdOrElseThrow(id);
         findPlan.setTitle(title);
         findPlan.setContents(contents);
 
@@ -65,6 +69,11 @@ public class PlanService {
 
     public void delete(Long id) {
         Plan findPlan = planRepository.findByIdOrElseThrow(id);
+        Comment findComment =  commentRepository.findByIdOrElseThrow(id);
+
+        if (!findComment.getComment().isEmpty()) {
+            throw new IllegalStateException("Plan에 연결된 댓글이 존재합니다. 삭제할 수 없습니다.");
+        }
 
         planRepository.delete(findPlan);
     }
